requirements.txt
google-generativeai>=0.8.0
fastapi>=0.115.0
mangum>=0.17.0
pydantic>=2.0.0
